backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
PyMuPDF==1.24.11
sentence-transformers==3.2.1
faiss-cpu==1.8.0.post1
python-dotenv==1.0.1
pytest==9.0.1
requests==2.32.3
pydantic==2.11.4
aiofiles==24.1.0
tqdm==4.66.6
httpx==0.28.1
rich==13.9.2
markdown==3.7
beautifulsoup4==4.12.3
redis==5.1.1
rq==1.16.2
python-multipart==0.0.17
openrouter==0.0.19
opentelemetry-api==1.27.0
sentry-sdk==2.18.0
